% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_numbers.R
\name{bd_branch_numbers}
\alias{bd_branch_numbers}
\title{Give branch numbering for BadiRate}
\usage{
bd_branch_numbers(tree, og = "./Orthogroups.GeneCount.csv",
  badirate_path = "", plot = FALSE, tree_file = "./branch_ids.tree")
}
\arguments{
\item{tree}{Ultrametric species tree.}

\item{og}{Orthogroup count file (Orthogroups.GeneCount.csv).}

\item{badirate_path}{Path to BadiRate.pl executable.}

\item{plot}{Logical. If TRUE, plots a tree with corresponding branch numbers. If FALSE, gives the tree in text.}

\item{tree_file}{Where to save the tree, which contains the branch ids.}
}
\value{
Either a plot or text of a labeled phylogenetic tree.
}
\description{
Give branch numbering for BadiRate
}
