% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_collect.R
\name{bd_collect}
\alias{bd_collect}
\title{Collect results files from BadiRate runs.}
\usage{
bd_collect(setup_table, out_dir = "./results")
}
\arguments{
\item{setup_table}{Setup table created when `prepare_badirate_scripts()` was invited.}

\item{out_dir}{Output directory to store result files.}
}
\value{
Null. Collects relevant informations in an organized manner.
}
\description{
Collect results files from BadiRate runs.
}
