% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_tree_ids.R
\name{bd_decode_tree_ids}
\alias{bd_decode_tree_ids}
\title{Decoding branch numbers to species compositions}
\usage{
bd_decode_tree_ids(tree_id)
}
\arguments{
\item{tree_id}{Ultrametric species tree with branch ids, generated with `bd_branch_numbers`}
}
\value{
A data_frame with 2 columns.
}
\description{
Decoding branch numbers to species compositions
}
