% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gains.R
\name{bd_extract_gain}
\alias{bd_extract_gain}
\title{Extract and parse gain/loss information by the selected best models}
\usage{
bd_extract_gain(best_models, results_dir = "./results")
}
\arguments{
\item{best_models}{Best models data frame generated by `bd_model_select()`}

\item{results_dir}{Directory where results was saved with `bd_collect()`.}
}
\value{
A dataframe with (orthogroups * branches) rows and 10 columns.
}
\description{
Extract and parse gain/loss information by the selected best models
}
