% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_wrappers.R
\name{qsub}
\alias{qsub}
\title{Wrapper to qsub command}
\usage{
qsub(files, ...)
}
\arguments{
\item{files}{Script files to be submitted.}

\item{...}{Other options to be passed to `qsub`.}
}
\description{
Wrapper to qsub command
}
