% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_wrappers.R
\name{sbatch}
\alias{sbatch}
\title{Wrapper to sbatch command}
\usage{
sbatch(files, ...)
}
\arguments{
\item{files}{Script files to be submitted.}

\item{...}{Other options to be passed to `sbatch`.}
}
\description{
Wrapper to sbatch command
}
