% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_orthogroups.R
\name{split_orthogroups}
\alias{split_orthogroups}
\title{Splitting and filtering orthogroups to individual files}
\usage{
split_orthogroups(og_file, max_count = 40,
  og_path = "./badirate_orthogroups")
}
\arguments{
\item{og_file}{Output from orthofinder (Orthogroups.GeneCount.csv) or similarly formatted count table.}

\item{og_path}{}
}
\value{
Null. Writes files to given path
}
\description{
Splitting and filtering orthogroups to individual files
}
